requirements.txt
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
cryptography @ file:///croot/cryptography_1673298753778/work
dnspython==2.3.0
email-validator==1.3.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
idna @ file:///croot/idna_1666125576474/work
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
toolz @ file:///croot/toolz_1667464077321/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
urllib3 @ file:///croot/urllib3_1673575502006/work
Werkzeug==2.2.2
WTForms==3.0.1
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
